% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stir.R
\name{convert.pec.sim.to.inbix}
\alias{convert.pec.sim.to.inbix}
\title{convert.pec.sim.to.inbix}
\usage{
convert.pec.sim.to.inbix(pEC.inputFile, inbix.file.prefix)
}
\arguments{
\item{pEC.inputFile}{pEC-simulated data csv file name with path if needed}

\item{inbix.file.prefix}{prefix of inbix filename for .num and .pheno files}
}
\value{
none
}
\description{
Reads a pEC simulated csv file, converts to inbix format and writes to working directory
}
\examples{
# setwd(...) # for output
infile <- "ARF_compare_1_multisurf_0.8_bias_No_k_0.1_pct.signals_1000_num.attr_100_num.samp.csv"
convert.pec.sim.to.inbix(infile,"simulated1")
# writes simulated1.num and simualted1.pheno

}
