% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reSTIR.R
\name{diffRegression}
\alias{diffRegression}
\title{diffRegression}
\usage{
diffRegression(pheno.diffs, predictor.diffs, regression.type = "lm")
}
\arguments{
\item{pheno.diffs}{outcome variable as vector of diffs}

\item{predictor.diffs}{one predictor varialbe as vector of diffs}

\item{regression.type}{(\code{"lm"}) todo: \code{"glm"} and covariates}
}
\value{
vector or regression stats to put into list for reSTIR and combine into matrix
}
\description{
Utility to run regression for a phenotype diff vector and one attribute diff vector. 
Organize regression statistics into a vector.
}
\examples{

}
