% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stir.R
\name{find.neighbors}
\alias{find.neighbors}
\title{find.neighbors}
\usage{
find.neighbors(attr.mat, pheno.class, method, k, delta.vec = NULL)
}
\arguments{
\item{test}{test}
}
\value{
test test
}
\description{
Description
}
\examples{
Example
}
