% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stir.R
\name{get.distance}
\alias{get.distance}
\title{get.distance}
\usage{
get.distance(attr.mat, metric)
}
\arguments{
\item{attr.mat}{m x p matrix of m instances and p attributes}

\item{metric}{for distance matrix between instances (\code{"manhattan"}, \code{"euclidean"}, ).}
}
\description{
Description
}
\examples{
Example
}
