% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mdd.RNAseq}
\alias{mdd.RNAseq}
\title{RNA-Seq from the paper referenced below.}
\format{Data frame with 157 mdd and hc subjects, 5912 genes.}
\usage{
data(mdd.RNAseq)
}
\description{
RNA-Seq from the paper referenced below.
}
\references{
Trang T. Le, et al. “Identification and replication of RNA-Seq gene network modules associated with depression severity,” Translational Psychiatry. 2018.
}
\keyword{datasets}
