% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stir.R
\name{regular.ttest.fn}
\alias{regular.ttest.fn}
\title{regular.ttest}
\usage{
regular.ttest.fn(attr.idx, dat, class.idx = ncol(dat))
}
\arguments{
\item{test}{test}
}
\value{
test test
}
\description{
Description
}
\examples{
Example
}
