% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reSTIR.R
\name{reSTIR}
\alias{reSTIR}
\title{reSTIR}
\usage{
reSTIR(pheno.vec, attr.mat, neighbor.pairs.idx,
  attr.diff.type = "manhattan", pheno.diff.type = "manhattan",
  regression.type = "linear")
}
\arguments{
\item{pheno.vec}{length-m numeric outcome vector for linear regression, factor for logistic regression}

\item{attr.mat}{m x p matrix of m instances and p attributes}

\item{neighbor.pairs.idx}{nearest hit/miss matrices, output from \code{find.neighbors}}

\item{attr.diff.type}{diff type for attributes (\code{"manhattan"} or \code{"euclidean"} for numeric)}

\item{pheno.diff.type}{diff type for phenotype (\code{"manhattan"} or \code{"euclidean"} for numeric)}

\item{regression.type}{(\code{"linear"} or \code{"logistic"})}
}
\value{
reSTIR.results.df: reSTIR regression coefficients and p-values for each attribute
}
\description{
regression-based STIR statistics for quantitative or case-control phenotypes
}
\examples{
neighbor.pairs.idx <- nearestNeighbors(predictors.mat, metric="manhattan", nbd.method = "multisurf", sd.frac = 0.5)
results.mat <- reSTIR(pheno.vec, predictors.mat, neighbor.pairs.idx, attr.diff.type="manhattan", pheno.diff.type="manhattan", regression.type="linear")

}
