% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stir.R
\name{stir}
\alias{stir}
\title{stir}
\usage{
stir(attr.mat, neighbor.idx, method, m = nrow(attr.mat), k, metric,
  transform = "None")
}
\arguments{
\item{attr.mat}{m x p matrix of m instances and p attributes}

\item{neighbor.idx}{nearest hit/miss matrix from \code{find.neighbors}}

\item{method}{neighborhood method [\code{"multisurf"} (k=0) or \code{"relieff"} (specify k)]}

\item{m}{optional number of instances}

\item{k}{number of nearest hits/misses for \code{"relieff"} method (k=0 for \code{"multisurf"})}

\item{metric}{for distance matrix between instances (\code{"manhattan"} or \code{"euclidean"})}

\item{transform}{transformation of distances (\code{"None"}, \code{"sqrt"} or \code{"neglog"})}
}
\value{
rs.list list: relief.score.df, arf.tstats.ordered, arf.fstats.ordered
}
\description{
Compute stir statistics for attributes given nearest hit/miss matrix
}
\examples{
#See vignette("STIRexample")
RF.method = "multisurf"
metric <- "manhattan"
neighbor.idx.observed <- find.neighbors(predictors.mat, pheno.class, k = 0, method = RF.method)
results.list <- stir(predictors.mat, neighbor.idx.observed, k = k, metric = metric, method = RF.method)
t_sorted_multisurf <- results.list$`STIR-t`
t_sorted_multisurf$attribute <- rownames(t_sorted_multisurf)

}
