% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reSTIR.R
\name{stirDistances}
\alias{stirDistances}
\title{stirDistances}
\usage{
stirDistances(attr.mat, metric = "manhattan")
}
\arguments{
\item{attr.mat}{m x p matrix of m instances and p attributes}

\item{metric}{for distance matrix between instances (default: \code{"manhattan"}, others include \code{"euclidean"}, 
versions scaled by max-min, \code{"relief-scaled-manhattan"} and \code{"relief-scaled-euclidean"}, and for GWAS \code{"allele-sharing-manhattan"}).}
}
\value{
distancesmat, matrix of m x m (instances x intances) pairwise distances.
}
\description{
Create m x m distance matrix from m instances and p attributes using different metrics. Used by nearestNeighbors(). 
Note: Probably best to standardize data before manhattan and euclidean.
}
\examples{
dist.mat <- stirDistances(predictors.mat, metric = "manhattan")
}
