% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reSTIR.R
\name{nearestNeighbors}
\alias{nearestNeighbors}
\title{nearestNeighbors}
\usage{
nearestNeighbors(attr.mat, metric = "manhattan",
  nbd.method = "multisurf", k = 0, sd.vec = NULL, sd.frac = 0.5)
}
\arguments{
\item{attr.mat}{m x p matrix of m instances and p attributes}

\item{metric}{used in stirDistances for distance matrix between instances, default: \code{"manhattan"} (numeric)}

\item{nbd.method}{neighborhood method [\code{"multisurf"} or \code{"surf"} (no k) or \code{"relieff"} (specify k)]}

\item{k}{number of constant nearest hits/misses for \code{"relieff"}}

\item{sd.frac}{multiplier of the standard deviation of the distances when subtracting from average for SURF or multiSURF.
The multiSURF default is sd.frac=0.5: mean - sd/2}
}
\value{
Ri_NN.idxmat, matrix of Ri's (first column) and their NN's (second column)
}
\description{
Find nearest neighbors of each instance using relief.method
Used for regression stir (reSTIR) (no hits or misses specified in function).
}
\examples{
#See vignette("STIRvignette")
nbd.method = "multisurf"
neighbor.pairs.idx <- nearestNeighbors(predictors.mat, metric="manhattan", nbd.method = nbd.method, sd.frac = 0.5)

}
