% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reSTIR.R
\name{stirDistances}
\alias{stirDistances}
\title{stirDistances}
\usage{
stirDistances(attr.mat, metric = "manhattan")
}
\arguments{
\item{attr.mat}{m x p matrix of m instances and p attributes}

\item{metric}{for distance matrix between instances (default: \code{"manhattan"}, \code{"euclidean"}, 
\code{"relief-scaled-manhattan"}, \code{"relief-scaled-euclidean"}, \code{"allele-sharing-manhattan"}).}
}
\description{
Description
}
\examples{
Example
}
